separate code to components, create context
