+++ src/components/ui/LandingCard.tsx
@@ -1,0 +1,12 @@
+export const LandingCard = () => {
+  return (
+    <header className="text-center">
+      <h1 className="text-5xl font-bold bg-clip-text text-transparent bg-gradient-to-r from-purple-400 via-pink-500 to-red-500">
+        Browser LLM
+      </h1>
+      <p className="text-gray-400 mt-2">
+        Run Large Language Models directly in your browser with WASM.
+      </p>
+    </header>
+  );
+};
